+++ Stripe/StripeiOSTests/STPApplePayFunctionalTest.swift
@@ -14,8 +14,8 @@ import XCTest
 @testable@_spi(STP) import StripeCore
 @testable@_spi(STP) import StripePayments
 @testable@_spi(STP) import StripePaymentSheet
-@testable@_spi(STP) import StripePaymentsUI
 @testable import StripePaymentsTestUtils
+@testable@_spi(STP) import StripePaymentsUI
 
 class STPApplePayFunctionalTest: STPNetworkStubbingTestCase {
     override func setUp() {
