+++ Stripe/StripeiOSTests/STPPaymentHandlerTests.swift
@@ -17,8 +17,8 @@ import XCTest
 @testable@_spi(STP) import StripeCore
 @testable@_spi(STP) import StripePayments
 @testable@_spi(STP) import StripePaymentSheet
-@testable@_spi(STP) import StripePaymentsUI
 @testable import StripePaymentsTestUtils
+@testable@_spi(STP) import StripePaymentsUI
 
 class STPPaymentHandlerStubbedTests: STPNetworkStubbingTestCase {
     override func setUp() {
