+++ Stripe/StripeiOSTests/STPPaymentMethodOptionsTest.swift
@@ -10,10 +10,10 @@ import XCTest
 
 @testable@_spi(STP) import Stripe
 @testable@_spi(STP) import StripeCore
+import StripeCoreTestUtils
 @testable@_spi(STP) import StripePayments
 @testable@_spi(STP) import StripePaymentSheet
 @testable@_spi(STP) import StripePaymentsUI
-import StripeCoreTestUtils
 
 class STPPaymentMethodOptionsTest: XCTestCase {
