+++ Tests/ReferenceImages_64/StripePaymentSheetTests.AddPaymentMethodViewControllerSnapshotTests/test_with_previous_customer_card_details_and_checkbox@3x.png
